Added types for hooks
