+++ react-components/src/index.ts
@@ -29,4 +29,5 @@ export type {
   AddReveal3DModelOptions
 } from './components/Reveal3DResources/types';
 export { RevealToolbar } from './components/RevealToolbar/RevealToolbar';
-export { useFdmAssetMappings, type FdmAssetMappingsConfig } from './hooks/useFdmAssetMappings';
+export { useFdmAssetMappings } from './hooks/useFdmAssetMappings';
+export { type FdmAssetMappingsConfig } from './hooks/types';
